Pass QWindow to PowerSaveBlocker.
