+++ Telegram/SourceFiles/calls/group/calls_group_panel.h
@@ -22,6 +22,10 @@ https://github.com/telegramdesktop/tdesktop/blob/master/LEGAL
 
 class Image;
 
+namespace base {
+class PowerSaveBlocker;
+} // namespace base
+
 namespace Data {
 class PhotoMedia;
 class CloudImageView;
@@ -197,6 +201,8 @@ private:
 #ifndef Q_OS_MAC
 	std::unique_ptr<Ui::Platform::SeparateTitleControls> _controls;
 #endif // !Q_OS_MAC
+
+	const std::unique_ptr<base::PowerSaveBlocker> _powerSaveBlocker;
 
 	rpl::lifetime _callLifetime;
